camel-jbang - Fixed tests

+++ dsl/camel-jbang/camel-jbang-it/src/test/java/org/apache/camel/dsl/jbang/it/support/JBangTestSupport.java
@@ -84,7 +84,9 @@ public abstract class JBangTestSupport {
         logger.debug("ending {}#{} using data folder {}", getClass().getName(), testInfo.getDisplayName(), getDataFolder());
         assertNoErrors();
         logger.debug("clean up data folder");
-        FileUtils.deleteQuietly(new File(containerDataFolder));
+        if (containerDataFolder != null) {
+            FileUtils.deleteQuietly(new File(containerDataFolder));
+        }
     }
 
     protected void stopAllRoutes() {
